1. move implic register func to ExtensionsRegister 2. move config to zipkin package
