+++ cmd/mosn/main/mosn.go
@@ -69,7 +69,6 @@ import (
 	_ "mosn.io/mosn/pkg/trace/sofa/http"
 	_ "mosn.io/mosn/pkg/trace/sofa/xprotocol"
 	_ "mosn.io/mosn/pkg/trace/sofa/xprotocol/bolt"
-	_ "mosn.io/mosn/pkg/trace/zipkin"
 	_ "mosn.io/mosn/pkg/upstream/healthcheck"
 	_ "mosn.io/mosn/pkg/upstream/servicediscovery/dubbod"
 	_ "mosn.io/mosn/pkg/wasm/abi/proxywasm010"
