+++ pkg/trace/zipkin/config.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package v2
+package zipkin
 
 import (
 	"errors"
